Feat(user): Implemented the user delete

+++ src/app/api/user/route.js
@@ -137,14 +137,17 @@ export async function DELETE(req) {
             return new Response(JSON.stringify({message: 'Internal Server Error: User not found.' }, { error: "User not found" }), { status: 404 });
         }
 
-        deletePhoto(existingUser.image.public_id)
+        // if user has an image, delete it from cloudinary
+        if (existingUser?.image){
+            deletePhoto(existingUser.image.public_id)
+        }
         
         // Find and delete the user by ID
         const deletedUser = await User.deleteOne({_id: id});
 
         // If no user is found, return a 404 response
         if (!deletedUser) {
-            return new Response(JSON.stringify({message: 'Internal Server Error: User not found.' }, { error: "Unable to delete user" }), { status: 500 });
+            return new Response(JSON.stringify({message: 'Internal Server Error: Unable to delete user.' }, { error: "Unable to delete user" }), { status: 500 });
         }
 
         // Return the deleted User as a response
